refactor BasePopupProvider & subclasses
This accomplishes the goal to sequester the in-flight request logic in the base
class.

The issues in the current code mostly came to the fact that it was insisting on
calling handlePermissionless from requestPermissions (and therefore having to
call `trackRequest` itself). This is not necessary: the function should just let
the base class know whether an approval is still needed after successfully
obtaining the extra permissions, and then the base class will call
handlePermissionless and track the request.

Doing this results in a really nice cleanup of the interfaces (no need to pass
resolve and reject down).